feat(plugins): plugin package basic implementation works
