+++ packages/polywrap-plugin/polywrap_plugin/package.py
@@ -1,9 +1,13 @@
-from typing import Generic
+from typing import Generic, Optional, cast
 
+from polywrap_core import IWrapPackage, Wrapper, GetManifestOptions
 from polywrap_manifest import AnyWrapManifest
-from polywrap_plugin import PluginModule, TConfig, TResult
+from polywrap_result import Ok, Result
 
-class PluginPackage(Generic[TConfig, TResult]):
+from .module import PluginModule, TConfig, TResult
+from .wrapper import PluginWrapper
+
+class PluginPackage(Generic[TConfig, TResult], IWrapPackage):
     module: PluginModule[TConfig, TResult]
     manifest: AnyWrapManifest
 
@@ -15,3 +19,8 @@ class PluginPackage(Generic[TConfig, TResult]):
         self.module = module
         self.manifest = manifest
 
+    async def create_wrapper(self) -> Result[Wrapper]:
+        return Ok(PluginWrapper(module=self.module, manifest=self.manifest))
+
+    async def get_manifest(self, options: Optional[GetManifestOptions] = None) -> Result[AnyWrapManifest]:
+        return await super().get_manifest(options)
