+++ packages/polywrap-plugin/tests/test_plugin_module.py
@@ -1,24 +1,15 @@
-from typing import Any, Dict
+from typing import Any
 
 import pytest
 from polywrap_client import PolywrapClient
-from polywrap_core import Invoker
 from polywrap_result import Ok
 
 from polywrap_plugin import PluginModule
 
-
-class GreetingModule(PluginModule[Any, str]):
-    def __init__(self, config: Any):
-        super().__init__(config)
-
-    def greeting(self, args: Dict[str, Any], client: Invoker):
-        return f"Greetings from: {args['name']}"
-
 @pytest.mark.asyncio
-async def test_plugin_module():
-    plugin = GreetingModule({})
+async def test_plugin_module(get_greeting_module: PluginModule[Any, Any]):
+    module = get_greeting_module
 
     client = PolywrapClient()
-    result = await plugin._wrap_invoke("greeting", { "name": "Joe" }, client) # type: ignore
+    result = await module._wrap_invoke("greeting", { "name": "Joe" }, client) # type: ignore
     assert result, Ok("Greetings from: Joe")
